+++ History.md
@@ -2,6 +2,7 @@ unreleased
 ==========
 
  * Accept string for `maxAge` (converted by `ms`)
+ * Add `headers` event
  * Include link in default redirect response
  * Use `EventEmitter.listenerCount` to count listeners